+++ doc/CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Highlighted Features
 
+* Fix request list page caching (Graeme Porteous)
 * Update locale cookie handling (Graeme Porteous)
 
 # 0.46.0.2